+++ R/StructuralVariants.R
@@ -1,0 +1,31 @@
+#' @name StructuralVariants
+#'
+#' @title Structural Variants
+#'
+#' @description  Concept examples of structural variants for test and
+#' demonstration purpose
+#'
+#' @details The variants represented here are the ones that can be represented
+#' in a [`GenomicBreaks`] object.  Short ones that are entirely contained in an
+#' aligned region, for instance as a gap, are not covered
+#'
+#' @format [`GBreaks`] objects described in more details in the
+#' _Structural Variants_ vignette, available with the command
+#' `vignette("Structural Variants", package = "GenomicBreaks")`.
+#'
+#' @author Charles Plessy
+#'
+#' @aliases exampleColinear exampleNotColinear exampleColinear3
+#' @aliases exampleInversion exampleTranslocation
+#' @aliases exampleInsertion exampleDeletion
+#'
+#' @family Structural variants
+#'
+#' @examples
+#' exampleColinear
+#' exampleNotColinear
+#' exampleColinear3
+#' exampleInversion
+#' exampleTranslocation
+#'
+NULL